==> Lessons/04_Lab_DataWrangling.Rmd <==
---
title: "Lab 4: Data Wrangling"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1- Answer questions on M3/A3
2- Answer questions on M4
3- Practice wrangling datasets with dplyr functions

## Set up your session

Today we will work with a dataset from the [North Temperate Lakes Long-Term Ecological Research Station](https://lter.limnology.wisc.edu/about/overview). The NTL-LTER is located in the boreal zone in northern Wisconsin, USA. We will use the [chemical and physical limnology dataset](https://lter.limnology.wisc.edu/content/cascade-project-north-temperate-lakes-lter-core-data-physical-and-chemical-limnology-1984), running from 1984-2016. 

Opening discussion: why might we be interested in long-term observations of temperature, oxygen, and light in lakes?

> Add notes here: 

```{r, message = FALSE}
getwd()
#install.packages(tidyverse)
library(tidyverse)
#install.packages(lubridate)
library(lubridate)
NTL.phys.data <- read.csv("./Data/Raw/NTL-LTER_Lake_ChemistryPhysics_Raw.csv", stringsAsFactors = TRUE)

class(NTL.phys.data$sampledate)
# Format sampledate as date
NTL.phys.data$sampledate <- as.Date(NTL.phys.data$sampledate, format = "%m/%d/%y")
```


## Filter

Filtering allows us to choose certain rows (observations) in our dataset.
When choosing rows, it is very important to know the class of the row. 

```{r}
class(NTL.phys.data$lakeid) #factor
class(NTL.phys.data$depth) #numeric

# dplyr filtering
#getting information about the lakes when the depth is 0 (surface)
#new data set at surface:
NTL.phys.data.surface <- filter(NTL.phys.data, depth == 0)

# Choose multiple conditions to filter
#Here, selecting by lake name:
summary(NTL.phys.data$lakename) #shows all possible lake names in dataset 
NTL.phys.data.PeterPaul <- filter(NTL.phys.data, lakename %in% c("Paul Lake", "Peter Lake")) 

# Choose a range of conditions of a numeric or integer variable
summary(NTL.phys.data$daynum)
NTL.phys.data.JunethruOctober <- filter(NTL.phys.data, daynum %in% c(152:304))
#here daynum represents June through October by creating a c() sequence

# Exercise 1: 
# filter NTL.phys.data for the year 1999
View(NTL.phys.data)
class(NTL.phys.data$year4) #class is an integer
NTL.phys.data.1999 <- filter(NTL.phys.data, year4 == 1999)
summary(NTL.phys.data.1999$year4) #yes all dates are 1999

# what code do you need to use, based on the class of the variable?
"=="

# Exercise 2: 
# filter NTL.phys.data for Tuesday Lake from 1990 through 1999.
NTL.phys.data.1990.1999 <- filter(NTL.phys.data, year4 %in% c(1990:1999))
summary(NTL.phys.data.1990.1999$year4) #checking that years have been selected 
```
Question: Why don't we filter using row numbers?

> Answer: 

## Pipes

Pipe is another method to wrangle datasets that looks cleaner and is easier to read.  We designate a pipe with `%>%`. A good way to think about the function of a pipe is with the word "then." 

Let's say we want to take our raw dataset (NTL.phys.data), *then* filter the data for Peter and Paul lakes, *then* select temperature and observation information, and *then* add a column for temperature in Fahrenheit: 

```{r}
NTL.phys.data.processed <- 
  NTL.phys.data %>% #choosing dataset
  filter(lakename == "Paul Lake" | lakename == "Peter Lake") %>% #filtering for lakes
  select(lakename, sampledate:temperature_C) %>% #selecting columns from sample date to temperature
  mutate(temperature_F = (temperature_C*9/5) + 32) #mutating the temperature 

#Exercise 3: Using a pipe filter NTL.phys.data for Tuesday Lake from 1990 through 1999 only for July 
NTL.phys.data$month <- strftime(NTL.phys.data$sampledate,"%m")
class(NTL.phys.data$month) #character class

NTL.phys.data.Tuesday.90s <- 
  NTL.phys.data %>% 
  filter(year4 %in% c(1990:1999)) %>% 
  filter(lakename == "Tuesday Lake") %>% 
  select(month == 07) #bc it is a character, using select instead of filter

#you can bypass the creation of a new month column:
#filter(month(sampledate))
  

#Exercise 4: Using the data from part 3, a pipe and the summarise() function find the mean surface temperature (hints: you will need to add another filter for depth==0). Make sure you eliminate NAs before computing the means

NTL.phys.data.Ex4 <- 
  NTL.phys.data.Tuesday.90s %>% 
  filter(depth == 0) %>% 
  filter(!is.na(temperature_C)) %>% 
  summarise(mean = mean(temperature_C))
#this creates only one value, which is the mean

```

## Gather and Spread

For gather we will use `pivot_longer` and for spread we will use `pivot_wider`.

```{r}
#Exercise 5: gather irradiance data (measured in the water column and measured on the deck of the sampling boat) into one column using pivot_longer


#Exercise 6: spread temperatureC into more than one column based on the depth

```
